Changes to allow the code to compile on the Arduino NANO 33 BLE
I've compiled these changes for every board I have installed and they seem fine. But there may be boards out there that do need the ```extern TwoWire Wire``` and ```extern SPIClass SPI```. Backward compatibility is not 100% guaranteed!

+++ src/ICM_20948.cpp
@@ -908,7 +908,7 @@ ICM_20948_Status_e ICM_20948_write_I2C(uint8_t reg, uint8_t *data, uint32_t len,
 
     _i2c->beginTransmission(addr);
     _i2c->write(reg);
-    _i2c->write(data, len);
+    _i2c->write(data, (uint8_t)len);
     _i2c->endTransmission();
 
     // for( uint32_t indi = 0; indi < len; indi++ ){
